refactor(fe): extract ProgressStep
